--- a/walk/walk.go
+++ b/walk/walk.go
@@ -65,36 +65,18 @@ func Descriptors(file protoreflect.FileDescriptor, fn func(protoreflect.Descript
 // The exit function is called using a post-order traversal, where the function
 // is called for a descriptor only after it is called for any descendants.
 func DescriptorsEnterAndExit(file protoreflect.FileDescriptor, enter, exit func(protoreflect.Descriptor) error) error {
-	for i := 0; i < file.Messages().Len(); i++ {
-		msg := file.Messages().Get(i)
-		if err := messageDescriptor(msg, enter, exit); err != nil {
-			return err
-		}
+	if err := walkContainer(file, enter, exit); err != nil {
+		return err
 	}
-	for i := 0; i < file.Enums().Len(); i++ {
-		en := file.Enums().Get(i)
-		if err := enumDescriptor(en, enter, exit); err != nil {
-			return err
-		}
-	}
-	for i := 0; i < file.Extensions().Len(); i++ {
-		ext := file.Extensions().Get(i)
-		if err := enter(ext); err != nil {
-			return err
-		}
-		if exit != nil {
-			if err := exit(ext); err != nil {
-				return err
-			}
-		}
-	}
-	for i := 0; i < file.Services().Len(); i++ {
-		svc := file.Services().Get(i)
+	services := file.Services()
+	for i, length := 0, services.Len(); i < length; i++ {
+		svc := services.Get(i)
 		if err := enter(svc); err != nil {
 			return err
 		}
-		for i := 0; i < svc.Methods().Len(); i++ {
-			mtd := svc.Methods().Get(i)
+		methods := svc.Methods()
+		for i, length := 0, methods.Len(); i < length; i++ {
+			mtd := methods.Get(i)
 			if err := enter(mtd); err != nil {
 				return err
 			}
@@ -113,12 +95,49 @@ func DescriptorsEnterAndExit(file protoreflect.FileDescriptor, enter, exit func(
 	return nil
 }
 
+type container interface {
+	Messages() protoreflect.MessageDescriptors
+	Enums() protoreflect.EnumDescriptors
+	Extensions() protoreflect.ExtensionDescriptors
+}
+
+func walkContainer(container container, enter, exit func(protoreflect.Descriptor) error) error {
+	messages := container.Messages()
+	for i, length := 0, messages.Len(); i < length; i++ {
+		msg := messages.Get(i)
+		if err := messageDescriptor(msg, enter, exit); err != nil {
+			return err
+		}
+	}
+	enums := container.Enums()
+	for i, length := 0, enums.Len(); i < length; i++ {
+		en := enums.Get(i)
+		if err := enumDescriptor(en, enter, exit); err != nil {
+			return err
+		}
+	}
+	exts := container.Extensions()
+	for i, length := 0, exts.Len(); i < length; i++ {
+		ext := exts.Get(i)
+		if err := enter(ext); err != nil {
+			return err
+		}
+		if exit != nil {
+			if err := exit(ext); err != nil {
+				return err
+			}
+		}
+	}
+	return nil
+}
+
 func messageDescriptor(msg protoreflect.MessageDescriptor, enter, exit func(protoreflect.Descriptor) error) error {
 	if err := enter(msg); err != nil {
 		return err
 	}
-	for i := 0; i < msg.Fields().Len(); i++ {
-		fld := msg.Fields().Get(i)
+	fields := msg.Fields()
+	for i, length := 0, fields.Len(); i < length; i++ {
+		fld := fields.Get(i)
 		if err := enter(fld); err != nil {
 			return err
 		}
@@ -128,8 +147,9 @@ func messageDescriptor(msg protoreflect.MessageDescriptor, enter, exit func(prot
 			}
 		}
 	}
-	for i := 0; i < msg.Oneofs().Len(); i++ {
-		oo := msg.Oneofs().Get(i)
+	oneofs := msg.Oneofs()
+	for i, length := 0, oneofs.Len(); i < length; i++ {
+		oo := oneofs.Get(i)
 		if err := enter(oo); err != nil {
 			return err
 		}
@@ -139,28 +159,8 @@ func messageDescriptor(msg protoreflect.MessageDescriptor, enter, exit func(prot
 			}
 		}
 	}
-	for i := 0; i < msg.Messages().Len(); i++ {
-		nested := msg.Messages().Get(i)
-		if err := messageDescriptor(nested, enter, exit); err != nil {
-			return err
-		}
-	}
-	for i := 0; i < msg.Enums().Len(); i++ {
-		en := msg.Enums().Get(i)
-		if err := enumDescriptor(en, enter, exit); err != nil {
-			return err
-		}
-	}
-	for i := 0; i < msg.Extensions().Len(); i++ {
-		ext := msg.Extensions().Get(i)
-		if err := enter(ext); err != nil {
-			return err
-		}
-		if exit != nil {
-			if err := exit(ext); err != nil {
-				return err
-			}
-		}
+	if err := walkContainer(msg, enter, exit); err != nil {
+		return err
 	}
 	if exit != nil {
 		if err := exit(msg); err != nil {
@@ -174,8 +174,9 @@ func enumDescriptor(en protoreflect.EnumDescriptor, enter, exit func(protoreflec
 	if err := enter(en); err != nil {
 		return err
 	}
-	for i := 0; i < en.Values().Len(); i++ {
-		enVal := en.Values().Get(i)
+	vals := en.Values()
+	for i, length := 0, vals.Len(); i < length; i++ {
+		enVal := vals.Get(i)
 		if err := enter(enVal); err != nil {
 			return err
 		}