I added the benchmarks, in the hopes of find a more efficient way to traverse descriptor protos, from `walk.DescriptorProtos`. The main cost is the way the fully-qualified names are computed/allocated as it traverses the descriptor hierarchy.

Ultimately, I did not come up with any better formulation. While I was able to reduce allocations slightly (from 316 allocs/op to 293 allocs/op), that approach actually yielded slightly worse throughput (consistently needing 5-7% more time per operation than the approach w/ slightly more allocations).

But _surprisingly_, the very simple changes to other walk function (`walk.Descriptors`), which just memoize the results of the various accessors, did actually result in a throughput improvement, consistently taking about 15% less time per operation.

Benchmark results below.

**before:**
```
BenchmarkDescriptors-10         	  352,322	      3,414 ns/op	       0 B/op	       0 allocs/op
BenchmarkDescriptorProtos-10    	  118,027	     10,055 ns/op	   16,232 B/op	     316 allocs/op
```

**after:**
```
BenchmarkDescriptors-10         	  409,730	      2,928 ns/op	       0 B/op	       0 allocs/op
BenchmarkDescriptorProtos-10    	  118,323	     10,008 ns/op	   16,232 B/op	     316 allocs/op
```